+++ src/libs/karm-gfx/path.cpp
@@ -65,14 +65,14 @@ void Path::_flattenQuadraticTo(Math::Vec2f start, Math::Vec2f cp, Math::Vec2f po
 
 void Path::_flattenArcTo(Math::Vec2f start, Math::Vec2f radius, double angle, Flags flags, Math::Vec2f point) {
     // Ported from canvg (https://code.google.com/p/canvg/)
-    float x1 = start.x;
-    float y1 = start.y;
-    float x2 = point.x;
-    float y2 = point.y;
+    double x1 = start.x;
+    double y1 = start.y;
+    double x2 = point.x;
+    double y2 = point.y;
 
-    float dx = x1 - x2;
-    float dy = y1 - y2;
-    float d = sqrtf(dx * dx + dy * dy);
+    double dx = x1 - x2;
+    double dy = y1 - y2;
+    double d = sqrtf(dx * dx + dy * dy);
 
     if (d < 1e-6f || radius.x < 1e-6f || radius.y < 1e-6f) {
         // The arc degenerates to a line
@@ -80,16 +80,16 @@ void Path::_flattenArcTo(Math::Vec2f start, Math::Vec2f radius, double angle, Fl
         return;
     }
 
-    float rotx = angle / 180.0f * M_PI; // x rotation angle
-    float sinrx = sinf(rotx);
-    float cosrx = cosf(rotx);
+    double rotx = angle / 180.0f * M_PI; // x rotation angle
+    double sinrx = sinf(rotx);
+    double cosrx = cosf(rotx);
 
     // Convert to center point parameterization.
     // http://www.w3.org/TR/SVG11/implnote.html#ArcImplementationNotes
 
     // 1) Compute x1', y1'
-    float x1p = cosrx * dx / 2.0f + sinrx * dy / 2.0f;
-    float y1p = -sinrx * dx / 2.0f + cosrx * dy / 2.0f;
+    double x1p = cosrx * dx / 2.0f + sinrx * dy / 2.0f;
+    double y1p = -sinrx * dx / 2.0f + cosrx * dy / 2.0f;
 
     d = Math::pow2(x1p) / Math::pow2(radius.x) + Math::pow2(y1p) / Math::pow2(radius.y);
 
@@ -100,14 +100,14 @@ void Path::_flattenArcTo(Math::Vec2f start, Math::Vec2f radius, double angle, Fl
     }
 
     // 2) Compute cx', cy'
-    float sa = Math::pow2(radius.x) * Math::pow2(radius.y) - Math::pow2(radius.x) * Math::pow2(y1p) - Math::pow2(radius.y) * Math::pow2(x1p);
-    float sb = Math::pow2(radius.x) * Math::pow2(y1p) + Math::pow2(radius.y) * Math::pow2(x1p);
+    double sa = Math::pow2(radius.x) * Math::pow2(radius.y) - Math::pow2(radius.x) * Math::pow2(y1p) - Math::pow2(radius.y) * Math::pow2(x1p);
+    double sb = Math::pow2(radius.x) * Math::pow2(y1p) + Math::pow2(radius.y) * Math::pow2(x1p);
 
     if (sa < 0.0f) {
         sa = 0.0f;
     }
 
-    float s = 0.0f;
+    double s = 0.0f;
 
     if (sb > 0.0f) {
         s = sqrtf(sa / sb);
@@ -120,19 +120,19 @@ void Path::_flattenArcTo(Math::Vec2f start, Math::Vec2f radius, double angle, Fl
         s = -s;
     }
 
-    float cxp = s * radius.x * y1p / radius.y;
-    float cyp = s * -radius.y * x1p / radius.x;
+    double cxp = s * radius.x * y1p / radius.y;
+    double cyp = s * -radius.y * x1p / radius.x;
 
     // 3) Compute cx,cy from cx',cy'
-    float cx = cosrx * cxp - sinrx * cyp + (x1 + x2) / 2.0f;
-    float cy = sinrx * cxp + cosrx * cyp + (y1 + y2) / 2.0f;
+    double cx = cosrx * cxp - sinrx * cyp + (x1 + x2) / 2.0f;
+    double cy = sinrx * cxp + cosrx * cyp + (y1 + y2) / 2.0f;
 
     // 4) Calculate theta1, and delta theta.
     Math::Vec2f u = {(x1p - cxp) / radius.x, (y1p - cyp) / radius.y};
     Math::Vec2f v = {(-x1p - cxp) / radius.x, (-y1p - cyp) / radius.y};
 
-    float a1 = Math::Vec2f(1, 0).angleWith(u); // Initial angle
-    float da = u.angleWith(v);
+    double a1 = Math::Vec2f(1, 0).angleWith(u); // Initial angle
+    double da = u.angleWith(v);
 
     if (!fs && da > 0) {
         da -= 2 * M_PI;
@@ -146,8 +146,8 @@ void Path::_flattenArcTo(Math::Vec2f start, Math::Vec2f radius, double angle, Fl
     // Split arc into max 90 degree segments.
     // The loop assumes an Iter per end point (including start and end), this +1.
     int ndivs = (int)(Math::abs(da) / (M_PI * 0.5f) + 1.0f);
-    float hda = (da / (float)ndivs) / 2.0f;
-    float kappa = Math::abs(4.0f / 3.0f * (1.0f - Math::cos(hda)) / Math::sin(hda));
+    double hda = (da / (double)ndivs) / 2.0f;
+    double kappa = Math::abs(4.0f / 3.0f * (1.0f - Math::cos(hda)) / Math::sin(hda));
 
     if (da < 0.0f) {
         kappa = -kappa;
@@ -157,7 +157,7 @@ void Path::_flattenArcTo(Math::Vec2f start, Math::Vec2f radius, double angle, Fl
     Math::Vec2f ptan = {};
 
     for (int i = 0; i <= ndivs; i++) {
-        float a = a1 + da * (i / (float)ndivs);
+        double a = a1 + da * (i / (double)ndivs);
 
         dx = Math::cos(a);
         dy = Math::sin(a);
